+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 * SharedUtils
+  * Add new `BraintreeSharedPreferencesException` to notify when an error occurs while interacting with shared preferences
   * Allow `BraintreeSharedPreferences` to gracefully degrade when `EncryptedSharedPreferences` fails (fix for #619)
 
 ## 4.19.0